Remove unnecessary layer/component
