+++ front/app/containers/IdeasShow/components/ProposalInfo/index.tsx
@@ -1,9 +1,32 @@
 import React from 'react';
 
+import {
+  Box,
+  colors,
+  fontSizes,
+  stylingConsts,
+  Text,
+} from '@citizenlab/cl2-component-library';
+import styled, { useTheme } from 'styled-components';
+
 import useIdeaStatus from 'api/idea_statuses/useIdeaStatus';
 import { IIdea } from 'api/ideas/types';
 
-import Status from './Status';
+import ReactionControl from 'components/ReactionControl';
+import T from 'components/T';
+
+import { ScreenReaderOnly } from 'utils/a11y';
+import { FormattedMessage } from 'utils/cl-intl';
+
+import CountDown from './components/CountDown';
+import ReactionCounter from './components/ReactionCounter';
+import messages from './messages';
+
+export const StatusHeading = styled(Text)`
+  font-size: ${fontSizes.base}px;
+  font-weight: bold;
+  text-transform: capitalize;
+`;
 
 interface Props {
   idea: IIdea;
@@ -11,6 +34,7 @@ interface Props {
 }
 
 const ProposalInfo = ({ idea, compact }: Props) => {
+  const theme = useTheme();
   const { data: ideaStatus } = useIdeaStatus(
     // TODO: Fix this the next time the file is edited.
     // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
@@ -19,8 +43,74 @@ const ProposalInfo = ({ idea, compact }: Props) => {
 
   if (!ideaStatus) return null;
 
+  const { code } = ideaStatus.data.attributes;
+
+  const showCountDown =
+    code === 'proposed' || code === 'expired' || code === 'custom';
+
+  const showProgressBar =
+    code === 'proposed' ||
+    code === 'threshold_reached' ||
+    code === 'custom' ||
+    code === 'ineligible' ||
+    code === 'answered';
+
+  const showVoteButtons =
+    code === 'proposed' ||
+    code === 'threshold_reached' ||
+    code === 'custom' ||
+    code === 'answered';
+
   return (
-    <Status idea={idea.data} ideaStatus={ideaStatus.data} compact={compact} />
+    <Box
+      display="flex"
+      flexDirection="column"
+      borderRadius={stylingConsts.borderRadius}
+    >
+      {showCountDown && (
+        <Box ml="auto">
+          <ScreenReaderOnly>
+            <FormattedMessage {...messages.a11y_timeLeft} />
+          </ScreenReaderOnly>
+          {idea.data.attributes.expires_at && (
+            <CountDown targetTime={idea.data.attributes.expires_at} />
+          )}
+        </Box>
+      )}
+      <Box display="flex" alignItems="center">
+        <Box
+          width="16px"
+          height="16px"
+          borderRadius={stylingConsts.borderRadius}
+          bg={ideaStatus.data.attributes.color}
+          mr="8px"
+        />
+        <StatusHeading>
+          <T value={ideaStatus.data.attributes.title_multiloc} />
+        </StatusHeading>
+      </Box>
+      <Box mb="24px" aria-live="polite">
+        <Text m="0">
+          <T value={ideaStatus.data.attributes.description_multiloc} />
+        </Text>
+      </Box>
+      {showProgressBar && (
+        <Box mb="24px">
+          <ReactionCounter
+            idea={idea.data}
+            barColor={theme.colors.tenantPrimary || colors.success}
+          />
+        </Box>
+      )}
+      {showVoteButtons && !compact && (
+        <ReactionControl
+          styleType="shadow"
+          ideaId={idea.data.id}
+          size="4"
+          variant="text"
+        />
+      )}
+    </Box>
   );
 };
